+++ example/AndroidOnlyExample/app/src/main/java/de/jensklingenberg/androidonlyexample/MainActivity.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.Modifier
 import androidx.lifecycle.lifecycleScope
 import de.jensklingenberg.androidonlyexample.ui.theme.AndroidOnlyExampleTheme
 import de.jensklingenberg.ktorfit.Ktorfit
-import de.jensklingenberg.ktorfit.adapter.FlowResponseConverter
+import de.jensklingenberg.ktorfit.converter.builtin.FlowResponseConverter
 import de.jensklingenberg.ktorfit.create
 import io.ktor.client.*
 import io.ktor.client.plugins.contentnegotiation.*
@@ -20,13 +20,11 @@ import io.ktor.serialization.kotlinx.json.*
 import kotlinx.coroutines.launch
 import kotlinx.serialization.json.Json
 
-val ktorfit = Ktorfit("https://swapi.dev/api/", HttpClient {
+val ktorfit = Ktorfit.Builder().baseUrl("https://swapi.dev/api/").httpClient(HttpClient {
     install(ContentNegotiation) {
         json(Json { isLenient = true; ignoreUnknownKeys = true })
     }
-}).also {
-    it.addResponseConverter(FlowResponseConverter())
-}
+}).responseConverter(FlowResponseConverter()).build()
 val api: StarWarsApi = ktorfit.create<StarWarsApi>()
 
 class MainActivity : ComponentActivity() {