+++ example/MultiplatformExample/shared/src/commonMain/kotlin/com/example/ktorfittest/StarWarsApi.kt
@@ -1,13 +1,17 @@
 package com.example.ktorfittest
 
+import de.jensklingenberg.ktorfit.Call
 import de.jensklingenberg.ktorfit.http.GET
 import de.jensklingenberg.ktorfit.http.Path
 import kotlinx.coroutines.flow.Flow
 
 interface StarWarsApi {
     @GET("people/{id}/")
-    suspend fun getPersonByIdResponse(@Path("id") peopleId: Int): People
+    suspend fun getPersonByIdResponse(@Path("id") peopleId: Int): Person
 
     @GET("people/{id}/")
-    fun getPeopleByIdFlowResponse(@Path("id") peopleId: Int): Flow<People>
+    fun getPeopleByIdFlowResponse(@Path("id") peopleId: Int): Flow<Person>
+
+    @GET("people/{id}/")
+    fun getPeopleByIdCallResponse(@Path("id") peopleId: Int): Call<Person>
 }