+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/parser/FunctionsParser.kt
@@ -2,23 +2,28 @@ package de.jensklingenberg.ktorfit.parser
 
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.symbol.KSFunctionDeclaration
-import de.jensklingenberg.ktorfit.*
-import de.jensklingenberg.ktorfit.model.KtorfitError
 import de.jensklingenberg.ktorfit.model.FunctionData
+import de.jensklingenberg.ktorfit.model.KtorfitError
 import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.BODY_PARAMETERS_CANNOT_BE_USED_WITH_FORM_OR_MULTI_PART_ENCODING
 import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.FIELD_MAP_PARAMETERS_CAN_ONLY_BE_USED_WITH_FORM_ENCODING
 import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.FIELD_PARAMETERS_CAN_ONLY_BE_USED_WITH_FORM_ENCODING
 import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.FORM_URL_ENCODED_CAN_ONLY_BE_SPECIFIED_ON_HTTP_METHODS_WITH_REQUEST_BODY
 import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.FOR_STREAMING_THE_RETURN_TYPE_MUST_BE_HTTP_STATEMENT
 import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.MISSING_EITHER_KEYWORD_URL_OrURL_PARAMETER
+import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.MISSING_X_IN_RELATIVE_URL_PATH
+import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.MULTIPART_CAN_ONLY_BE_SPECIFIED_ON_HTTPMETHODS
 import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.MULTIPLE_URL_METHOD_ANNOTATIONS_FOUND
 import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.NON_BODY_HTTP_METHOD_CANNOT_CONTAIN_BODY
+import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.NO_HTTP_ANNOTATION_AT
 import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.ONLY_ONE_ENCODING_ANNOTATION_IS_ALLOWED
 import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.ONLY_ONE_HTTP_METHOD_IS_ALLOWED
 import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.ONLY_ONE_REQUEST_BUILDER_IS_ALLOWED
 import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.PATH_CAN_ONLY_BE_USED_WITH_RELATIVE_URL_ON
+import de.jensklingenberg.ktorfit.model.KtorfitError.Companion.URL_CAN_ONLY_BE_USED_WITH_EMPY
 import de.jensklingenberg.ktorfit.model.TypeData
 import de.jensklingenberg.ktorfit.model.annotations.*
+import de.jensklingenberg.ktorfit.model.ktorfitError
+import de.jensklingenberg.ktorfit.utils.*
 
 
 fun getHttpMethodAnnotations(func: KSFunctionDeclaration): List<HttpMethodAnnotation> {
@@ -93,11 +98,11 @@ fun getFunctionDataList(
         val httpMethodAnnoList = getHttpMethodAnnotations(funcDeclaration)
 
         if (httpMethodAnnoList.isEmpty()) {
-            logger.ktorfitError("No Http annotation $functionName", funcDeclaration)
+            logger.ktorfitError(NO_HTTP_ANNOTATION_AT(functionName), funcDeclaration)
         }
 
         if (httpMethodAnnoList.size > 1) {
-            logger.ktorfitError(ONLY_ONE_HTTP_METHOD_IS_ALLOWED+ "Found: " + httpMethodAnnoList.joinToString { it.httpMethod.keyword } + " at " + functionName,
+            logger.ktorfitError(ONLY_ONE_HTTP_METHOD_IS_ALLOWED + "Found: " + httpMethodAnnoList.joinToString { it.httpMethod.keyword } + " at " + functionName,
                 funcDeclaration)
         }
 
@@ -118,15 +123,15 @@ fun getFunctionDataList(
         when (httpMethodAnno.httpMethod) {
             HttpMethod.POST, HttpMethod.PUT, HttpMethod.PATCH -> {}
             else -> {
-                if(httpMethodAnno is CustomHttp && httpMethodAnno.hasBody){
-                 //Do nothing
-                }else if (functionParameters.any { it.hasAnnotation<Body>() }) {
+                if (httpMethodAnno is CustomHttp && httpMethodAnno.hasBody) {
+                    //Do nothing
+                } else if (functionParameters.any { it.hasAnnotation<Body>() }) {
                     logger.ktorfitError(NON_BODY_HTTP_METHOD_CANNOT_CONTAIN_BODY, funcDeclaration)
                 }
 
                 if (functionAnnotationList.any { it is Multipart }) {
                     logger.ktorfitError(
-                        "Multipart can only be specified on HTTP methods with request body (e.g., @POST)",
+                        MULTIPART_CAN_ONLY_BE_SPECIFIED_ON_HTTPMETHODS,
                         funcDeclaration
                     )
                 }
@@ -142,9 +147,19 @@ fun getFunctionDataList(
 
         if (functionParameters.any { it.hasAnnotation<Path>() } && httpMethodAnno.path.isEmpty()) {
             logger.ktorfitError(
-                PATH_CAN_ONLY_BE_USED_WITH_RELATIVE_URL_ON+"@${httpMethodAnno.httpMethod.keyword}",
+                PATH_CAN_ONLY_BE_USED_WITH_RELATIVE_URL_ON + "@${httpMethodAnno.httpMethod.keyword}",
                 funcDeclaration
             )
+        }
+
+        functionParameters.filter { it.hasAnnotation<Path>() }.forEach {
+            val pathAnnotation = it.findAnnotationOrNull<Path>()
+            if (!httpMethodAnno.path.contains("{${pathAnnotation?.value ?: ""}}")) {
+                logger.ktorfitError(
+                    MISSING_X_IN_RELATIVE_URL_PATH(pathAnnotation?.value ?: ""),
+                    funcDeclaration
+                )
+            }
         }
 
         if (funcDeclaration.getFormUrlEncodedAnnotation() != null && funcDeclaration.getMultipartAnnotation() != null) {
@@ -157,28 +172,22 @@ fun getFunctionDataList(
             }
             if (httpMethodAnno.path.isNotEmpty()) {
                 logger.ktorfitError(
-                    "@Url only be used with empty @${httpMethodAnno.httpMethod.keyword} URL value",
+                    URL_CAN_ONLY_BE_USED_WITH_EMPY(httpMethodAnno.httpMethod.keyword),
                     funcDeclaration
                 )
             }
         }
 
-        if (functionParameters.any { it.hasAnnotation<Field>() }) {
-            if (funcDeclaration.getFormUrlEncodedAnnotation() == null) {
-                logger.ktorfitError(FIELD_PARAMETERS_CAN_ONLY_BE_USED_WITH_FORM_ENCODING, funcDeclaration)
-            }
+        if (functionParameters.any { it.hasAnnotation<Field>() } && funcDeclaration.getFormUrlEncodedAnnotation() == null) {
+            logger.ktorfitError(FIELD_PARAMETERS_CAN_ONLY_BE_USED_WITH_FORM_ENCODING, funcDeclaration)
         }
 
-        if (functionParameters.any { it.hasAnnotation<FieldMap>() }) {
-            if (funcDeclaration.getFormUrlEncodedAnnotation() == null) {
-                logger.ktorfitError(FIELD_MAP_PARAMETERS_CAN_ONLY_BE_USED_WITH_FORM_ENCODING, funcDeclaration)
-            }
+        if (functionParameters.any { it.hasAnnotation<FieldMap>() } && funcDeclaration.getFormUrlEncodedAnnotation() == null) {
+            logger.ktorfitError(FIELD_MAP_PARAMETERS_CAN_ONLY_BE_USED_WITH_FORM_ENCODING, funcDeclaration)
         }
 
-        if (functionParameters.any { it.hasAnnotation<Body>() }) {
-            if (funcDeclaration.getFormUrlEncodedAnnotation() != null) {
-                logger.ktorfitError(BODY_PARAMETERS_CANNOT_BE_USED_WITH_FORM_OR_MULTI_PART_ENCODING, funcDeclaration)
-            }
+        if (functionParameters.any { it.hasAnnotation<Body>() } && funcDeclaration.getFormUrlEncodedAnnotation() != null) {
+            logger.ktorfitError(BODY_PARAMETERS_CANNOT_BE_USED_WITH_FORM_OR_MULTI_PART_ENCODING, funcDeclaration)
         }
 
         return@map FunctionData(