+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/requestData/GetBodyDataText.kt
@@ -1,0 +1,10 @@
+package de.jensklingenberg.ktorfit.requestData
+
+import de.jensklingenberg.ktorfit.model.ParameterData
+import de.jensklingenberg.ktorfit.model.annotations.Body
+
+fun getBodyDataText(params: List<ParameterData>): String {
+    return params.firstOrNull { it.hasAnnotation<Body>() }?.let {
+        "bodyData = " + it.name
+    } ?: ""
+}