+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/requestData/HeadersArgumentText.kt
@@ -3,6 +3,7 @@ package de.jensklingenberg.ktorfit.requestData
 import de.jensklingenberg.ktorfit.*
 import de.jensklingenberg.ktorfit.model.ParameterData
 import de.jensklingenberg.ktorfit.model.annotations.*
+import de.jensklingenberg.ktorfit.utils.surroundIfNotEmpty
 
 
 /**
@@ -11,7 +12,7 @@ import de.jensklingenberg.ktorfit.model.annotations.*
 
 fun getHeadersArgumentText(
     functionAnnotations: List<FunctionAnnotation>,
-     paramList: List<ParameterData>
+    paramList: List<ParameterData>
 ): String {
     val headerList = mutableListOf<Pair<String, String>>()
 
@@ -23,7 +24,7 @@ fun getHeadersArgumentText(
         /**
          * Can't add Content Type Header, because it leads to Ktor issues https://github.com/ktorio/ktor/issues/1127
          */
-      //  headerList.add(Pair("\"Content-Type\"", "\"application/x-www-form-urlencoded\""))
+        //  headerList.add(Pair("\"Content-Type\"", "\"application/x-www-form-urlencoded\""))
     }
 
     paramsWithHeaderAnno.forEach { myParam ->
@@ -50,8 +51,8 @@ fun getHeadersArgumentText(
         val (key, value) = it
 
         "HeaderData($key,$value)"
-    }.surroundIfNotEmpty("listOf(", ")")
+    }.surroundIfNotEmpty("headers = listOf(", ")")
 
 
-    return headerText.prefixIfNotEmpty("headers = ").replace(" ","·")
+    return headerText.replace(" ", "·")
 }