+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/requestData/RequestDataArgumentNode.kt
@@ -1,24 +1,9 @@
 package de.jensklingenberg.ktorfit.requestData
 
-import de.jensklingenberg.ktorfit.hasAnnotation
+
+import de.jensklingenberg.ktorfit.generator.requestDataClass
 import de.jensklingenberg.ktorfit.model.FunctionData
-import de.jensklingenberg.ktorfit.model.ParameterData
-import de.jensklingenberg.ktorfit.model.annotations.Body
 
-
-private fun getRequestBuilderText(parameterDataList: List<ParameterData>): String {
-    return (parameterDataList.find { it.hasRequestBuilderAnno }?.let {
-        "requestBuilder = " + it.name
-    } ?: "")
-}
-
-private fun getBodyDataText(params: List<ParameterData>): String {
-    var bodyText = ""
-    params.firstOrNull { it.hasAnnotation<Body>() }?.let {
-        bodyText = "bodyData = " + it.name
-    }
-    return bodyText
-}
 
 fun getRequestDataArgumentText(functionData: FunctionData): String {
     val methodAnnotation = functionData.httpMethodAnnotation
@@ -30,6 +15,7 @@ fun getRequestDataArgumentText(functionData: FunctionData): String {
     val body = getBodyDataText(functionData.parameterDataList)
     //URL
     val urlPath = getRelativeUrlArgumentText(functionData.httpMethodAnnotation, functionData.parameterDataList)
+    val pathsText = getPathsText(functionData.parameterDataList)
     val queryText = getQueryArgumentText(functionData.parameterDataList)
     val fieldsText = getFieldArgumentsText(functionData.parameterDataList)
     val partsText = getPartsArgumentText(functionData.parameterDataList)
@@ -46,8 +32,9 @@ fun getRequestDataArgumentText(functionData: FunctionData): String {
         fieldsText,
         partsText,
         builderText,
-        qualifiedTypeName
+        qualifiedTypeName,
+        pathsText
     ).filter { it.isNotEmpty() }.joinToString(",\n") { it }
 
-    return "val requestData = RequestData($args) \n"
+    return "val requestData = ${requestDataClass.name}($args) \n"
 }