+++ ktorfit-lib/src/commonMain/kotlin/de/jensklingenberg/ktorfit/Ktorfit.kt
@@ -1,42 +1,136 @@
 package de.jensklingenberg.ktorfit
 
-import de.jensklingenberg.ktorfit.adapter.KtorfitCallResponseConverter
-import de.jensklingenberg.ktorfit.adapter.ResponseConverter
+import de.jensklingenberg.ktorfit.converter.CoreResponseConverter
+import de.jensklingenberg.ktorfit.converter.ResponseConverter
+import de.jensklingenberg.ktorfit.converter.SuspendResponseConverter
 import io.ktor.client.*
+import io.ktor.client.engine.*
 
 
 /**
  * Main class for Ktorfit, create the class than use the [create<T>()] function.
- * @param baseUrl that will be used for every request with object
- * @param httpClient client that will be used for every request with object
  */
-class Ktorfit(var baseUrl: String, var httpClient: HttpClient = HttpClient()) {
+class Ktorfit private constructor(
+    val baseUrl: String,
+    val httpClient: HttpClient = HttpClient(),
+    val responseConverters: Set<ResponseConverter>,
+    val suspendResponseConverters: Set<SuspendResponseConverter>
+) {
 
-    private val responseConverters: MutableList<ResponseConverter> = mutableListOf()
-
-
-    init {
-        if (baseUrl.isEmpty()) {
-            throw IllegalStateException("Base URL required")
-        }
-
-        if (!baseUrl.endsWith("/")) {
-            throw IllegalStateException("Base URL needs to end with /")
-        }
-        addResponseConverter(KtorfitCallResponseConverter())
-    }
+    @Deprecated(
+        "Use Ktorfit.Builder()",
+        replaceWith = ReplaceWith("Ktorfit.Builder().baseUrl(baseUrl).httpClient(httpClient).build()")
+    )
+    constructor(
+        baseUrl: String,
+        httpClient: HttpClient = HttpClient()
+    ) : this(baseUrl, httpClient, emptySet(), emptySet())
 
     /**
-     * Use this to add [ResponseConverter] for unsupported return types of requests
+     * Builder class for Ktorfit.
+     *
+     * @see baseUrl
+     * @see httpClient
      */
-    fun addResponseConverter(ResponseConverter: ResponseConverter): Ktorfit {
-        responseConverters.add(ResponseConverter)
-        return this
+    class Builder {
+        private var _baseUrl: String = ""
+        private var _httpClient = HttpClient()
+        private var _responseConverter: MutableSet<ResponseConverter> = mutableSetOf()
+        private var _suspendResponseConverter: MutableSet<SuspendResponseConverter> = mutableSetOf()
+
+        /**
+         * That will be used for every request with object
+         */
+        fun baseUrl(url: String) = apply {
+            if (url.isEmpty()) {
+                throw IllegalStateException("Base URL required")
+            }
+
+            if (!url.endsWith("/")) {
+                throw IllegalStateException("Base URL needs to end with /")
+            }
+            this._baseUrl = url
+        }
+
+        /**
+         * Client that will be used for every request with object
+         */
+        fun httpClient(client: HttpClient) = apply {
+            this._httpClient = client
+        }
+
+        /**
+         * Build HttpClient by just passing an engine
+         */
+        fun httpClient(engine: HttpClientEngine) = apply {
+            this._httpClient = HttpClient(engine)
+        }
+
+        /**
+         * Build HttpClient by just passing an engine factory
+         */
+        fun <T : HttpClientEngineConfig> httpClient(engineFactory: HttpClientEngineFactory<T>) = apply {
+            this._httpClient = HttpClient(engineFactory)
+        }
+
+        /**
+         * Client-Builder that will be used for every request with object
+         */
+        fun httpClient(config: HttpClientConfig<*>.() -> Unit) = apply {
+            this._httpClient = HttpClient(this._httpClient.engine, config)
+        }
+
+        /**
+         * Client-Builder with engine that will be used for every request with object
+         */
+        fun httpClient(engine: HttpClientEngine, config: HttpClientConfig<*>.() -> Unit) = apply {
+            this._httpClient = HttpClient(engine, config)
+        }
+
+        /**
+         * Client-Builder with engine factory that will be used for every request with object
+         */
+        fun <T : HttpClientEngineConfig> httpClient(engineFactory: HttpClientEngineFactory<T>, config: HttpClientConfig<T>.() -> Unit) = apply {
+            this._httpClient = HttpClient(engineFactory, config)
+        }
+
+        /**
+         * Use this to add [ResponseConverter] or [SuspendResponseConverter] for unsupported return types of requests
+         */
+        fun responseConverter(vararg converters: CoreResponseConverter) = apply {
+            converters.forEach { converter ->
+                if (converter is ResponseConverter) {
+                    this._responseConverter.add(converter)
+                }
+                if (converter is SuspendResponseConverter) {
+                    this._suspendResponseConverter.add(converter)
+                }
+            }
+        }
+
+        /**
+         * Apply changes to builder and get the Ktorfit instance without the need of calling [build] afterwards.
+         */
+        fun build(builder: Builder.() -> Unit) = this.apply(builder).build()
+
+        /**
+         * Creates an instance of Ktorfit with specified baseUrl and HttpClient.
+         */
+        fun build(): Ktorfit {
+            return Ktorfit(_baseUrl, _httpClient, _responseConverter, _suspendResponseConverter)
+        }
     }
-
-    fun getResponseConverters() = responseConverters
-
 }
+
+/**
+ * Create a Ktorfit instance using Kotlin-DSL.
+ */
+fun ktorfit(builder: Ktorfit.Builder.() -> Unit) = Ktorfit.Builder().apply(builder).build()
+
+/**
+ * Creates a Ktorfit Builder instance using Kotlin-DSL.
+ */
+fun ktorfitBuilder(builder: Ktorfit.Builder.() -> Unit) = Ktorfit.Builder().apply(builder)
 
 /**
  * This will make IntelliJ think that this function exists.