+++ ktorfit-lib/src/commonMain/kotlin/de/jensklingenberg/ktorfit/converter/builtin/KtorfitCallResponseConverter.kt
@@ -1,7 +1,8 @@
-package de.jensklingenberg.ktorfit.adapter
+package de.jensklingenberg.ktorfit.converter.builtin
 
 import de.jensklingenberg.ktorfit.Call
 import de.jensklingenberg.ktorfit.Callback
+import de.jensklingenberg.ktorfit.converter.ResponseConverter
 import io.ktor.client.statement.*
 import io.ktor.util.reflect.*
 import kotlinx.coroutines.GlobalScope
@@ -10,10 +11,11 @@ import kotlinx.coroutines.launch
 
 /**
  * Converter to enable the use of Call<> as return type
+ * e.g. fun test(): Call<String>
  */
 class KtorfitCallResponseConverter : ResponseConverter {
 
-    override fun supportedType(returnTypeName: String): Boolean {
+    override fun supportedType(returnTypeName: String, isSuspend: Boolean): Boolean {
         return returnTypeName == "de.jensklingenberg.ktorfit.Call"
     }