+++ ktorfit-lib/src/commonMain/kotlin/de/jensklingenberg/ktorfit/internal/HeaderData.kt
@@ -1,3 +1,5 @@
 package de.jensklingenberg.ktorfit.internal
 
 data class HeaderData(val key: String, val value: Any?)
+
+data class PathData(val key: String, val value: String, val encoded: Boolean)