+++ ktorfit-lib/src/jvmTest/kotlin/QueryTest.kt
@@ -25,13 +25,13 @@ class QueryTest {
             }
         }
 
-        val ktorfit = Ktorfit("www.test.de/", HttpClient(engine))
+        val ktorfit = Ktorfit.Builder().baseUrl("www.test.de/").httpClient(HttpClient(engine)).build()
         runBlocking {
             val requestData = RequestData(
                 method = "GET",
                 relativeUrl = "",
                 qualifiedRawTypeName = "kotlin.String",
-                queries = listOf(QueryData(false, testKey, testValue, QueryType.QUERY))
+                queries = listOf(QueryData(testKey, testValue, false, QueryType.QUERY))
             )
             KtorfitClient(ktorfit).suspendRequest<String, String>(requestData)
         }
@@ -49,13 +49,13 @@ class QueryTest {
             }
         }
 
-        val ktorfit = Ktorfit("www.test.de/", HttpClient(engine))
+        val ktorfit = Ktorfit.Builder().baseUrl("www.test.de/").httpClient(HttpClient(engine)).build()
         runBlocking {
             val requestData = RequestData(
                 method = "GET",
                 relativeUrl = "",
                 qualifiedRawTypeName = "kotlin.String",
-                queries = listOf(QueryData(true, testKey, testValue, QueryType.QUERY))
+                queries = listOf(QueryData(testKey, testValue, true, QueryType.QUERY))
             )
             KtorfitClient(ktorfit).suspendRequest<String, String>(requestData)
         }
@@ -75,13 +75,13 @@ class QueryTest {
             }
         }
 
-        val ktorfit = Ktorfit("www.test.de/", HttpClient(engine))
+        val ktorfit = Ktorfit.Builder().baseUrl("www.test.de/").httpClient(HttpClient(engine)).build()
         runBlocking {
             val requestData = RequestData(
                 method = "GET",
                 relativeUrl = "",
                 qualifiedRawTypeName = "kotlin.String",
-                queries = listOf(QueryData(false, testKey, testValue, QueryType.QUERY))
+                queries = listOf(QueryData(testKey, testValue, false, QueryType.QUERY))
             )
             KtorfitClient(ktorfit).suspendRequest<String, String>(requestData)
         }
@@ -101,13 +101,13 @@ class QueryTest {
             }
         }
 
-        val ktorfit = Ktorfit("www.test.de/", HttpClient(engine))
+        val ktorfit = Ktorfit.Builder().baseUrl("www.test.de/").httpClient(HttpClient(engine)).build()
         runBlocking {
             val requestData = RequestData(
                 method = "GET",
                 relativeUrl = "",
                 qualifiedRawTypeName = "kotlin.String",
-                queries = listOf(QueryData(false, testKey, testValue, QueryType.QUERY))
+                queries = listOf(QueryData(testKey, testValue, false, QueryType.QUERY))
             )
             KtorfitClient(ktorfit).suspendRequest<String, String>(requestData)
         }
@@ -126,13 +126,13 @@ class QueryTest {
             }
         }
 
-        val ktorfit = Ktorfit("www.test.de/", HttpClient(engine))
+        val ktorfit = Ktorfit.Builder().baseUrl("www.test.de/").httpClient(HttpClient(engine)).build()
         runBlocking {
             val requestData = RequestData(
                 method = "GET",
                 relativeUrl = "",
                 qualifiedRawTypeName = "kotlin.String",
-                queries = listOf(QueryData(false, testKey, testValue, QueryType.QUERYNAME))
+                queries = listOf(QueryData(testKey, testValue, false, QueryType.QUERYNAME))
             )
             KtorfitClient(ktorfit).suspendRequest<String, String>(requestData)
         }
@@ -150,13 +150,13 @@ class QueryTest {
             }
         }
 
-        val ktorfit = Ktorfit("www.test.de/", HttpClient(engine))
+        val ktorfit = Ktorfit.Builder().baseUrl("www.test.de/").httpClient(HttpClient(engine)).build()
         runBlocking {
             val requestData = RequestData(
                 method = "GET",
                 relativeUrl = "",
                 qualifiedRawTypeName = "kotlin.String",
-                queries = listOf(QueryData(false, testKey, testValue, QueryType.QUERYNAME))
+                queries = listOf(QueryData(testKey, testValue, false, QueryType.QUERYNAME))
             )
             KtorfitClient(ktorfit).suspendRequest<String, String>(requestData)
         }
@@ -174,13 +174,13 @@ class QueryTest {
             }
         }
 
-        val ktorfit = Ktorfit("www.test.de/", HttpClient(engine))
+        val ktorfit = Ktorfit.Builder().baseUrl("www.test.de/").httpClient(HttpClient(engine)).build()
         runBlocking {
             val requestData = RequestData(
                 method = "GET",
                 relativeUrl = "",
                 qualifiedRawTypeName = "kotlin.String",
-                queries = listOf(QueryData(false, testKey, testMap, QueryType.QUERYMAP))
+                queries = listOf(QueryData(testKey, testMap, false, QueryType.QUERYMAP))
             )
             KtorfitClient(ktorfit).suspendRequest<String, String>(requestData)
         }