+++ sandbox/src/jvmMain/kotlin/de/jensklingenberg/ktorfit/demo/CreateIssue.kt
@@ -1,12 +1,11 @@
 package de.jensklingenberg.ktorfit.demo
 
 import com.example.api.GithubService
-import de.jensklingenberg.ktorfit.adapter.FlowResponseConverter
-import de.jensklingenberg.ktorfit.Ktorfit
-import de.jensklingenberg.ktorfit.adapter.KtorfitCallResponseConverter
+import de.jensklingenberg.ktorfit.converter.builtin.FlowResponseConverter
+import de.jensklingenberg.ktorfit.converter.builtin.KtorfitSuspendCallResponseConverter
 import de.jensklingenberg.ktorfit.create
+import de.jensklingenberg.ktorfit.ktorfit
 import io.ktor.client.*
-import io.ktor.client.plugins.*
 import io.ktor.client.plugins.contentnegotiation.*
 import io.ktor.serialization.kotlinx.json.*
 import kotlinx.coroutines.delay
@@ -27,15 +26,17 @@ fun main() {
     }
 
 
-    val jvmKtorfit = Ktorfit(baseUrl = GithubService.baseUrl, jvmClient)
+    val jvmKtorfit = ktorfit {
+        baseUrl(GithubService.baseUrl)
+        httpClient(jvmClient)
+        responseConverter(
+            FlowResponseConverter(),
+            RxResponseConverter(),
+            KtorfitSuspendCallResponseConverter()
+        )
+    }
 
     val testApi = jvmKtorfit.create<GithubService>()
-
-
-
-    jvmKtorfit.addResponseConverter(FlowResponseConverter())
-    jvmKtorfit.addResponseConverter(RxResponseConverter())
-    jvmKtorfit.addResponseConverter(KtorfitCallResponseConverter())
 
 
     runBlocking {