+++ sandbox/src/jvmMain/kotlin/de/jensklingenberg/ktorfit/demo/JvMMain.kt
@@ -2,10 +2,10 @@ package de.jensklingenberg.ktorfit.demo
 
 
 import com.example.api.StarWarsApi
-import de.jensklingenberg.ktorfit.Ktorfit
-import de.jensklingenberg.ktorfit.adapter.FlowResponseConverter
-import de.jensklingenberg.ktorfit.adapter.KtorfitCallResponseConverter
+import de.jensklingenberg.ktorfit.converter.builtin.FlowResponseConverter
+import de.jensklingenberg.ktorfit.converter.builtin.KtorfitSuspendCallResponseConverter
 import de.jensklingenberg.ktorfit.create
+import de.jensklingenberg.ktorfit.ktorfit
 import io.ktor.client.*
 import io.ktor.client.plugins.contentnegotiation.*
 import io.ktor.serialization.kotlinx.json.*
@@ -25,25 +25,24 @@ val jvmClient = HttpClient {
 
 }
 
-val jvmKtorfit = Ktorfit(baseUrl = StarWarsApi.baseUrl, jvmClient)
+val jvmKtorfit = ktorfit {
+    baseUrl(StarWarsApi.baseUrl)
+    httpClient(jvmClient)
+    responseConverter(
+        FlowResponseConverter(),
+        RxResponseConverter(),
+        KtorfitSuspendCallResponseConverter(),
+        SuspendConverter()
+    )
+}
 
 
 fun main() {
-
-    jvmKtorfit.addResponseConverter(FlowResponseConverter())
-    jvmKtorfit.addResponseConverter(RxResponseConverter())
-    jvmKtorfit.addResponseConverter(KtorfitCallResponseConverter())
-
-
-    val exampleApi = jvmKtorfit.create<JvmPlaceHolderApi>()
-
-
-
-
+    val exampleApi = jvmKtorfit.createJvmPlaceHolderApi()
 
     println("==============================================")
     runBlocking {
-        val response = exampleApi.getPersonById2(1)
+        val response = exampleApi.getPersonById2(2)
 
         println("LI    " + response)