+++ sandbox/src/linuxX64Main/kotlin/LinuxMain.kt
@@ -1,17 +1,18 @@
 import com.example.api.JsonPlaceHolderApi
 import de.jensklingenberg.ktorfit.Ktorfit
-import de.jensklingenberg.ktorfit.adapter.FlowResponseConverter
+import de.jensklingenberg.ktorfit.converter.builtin.FlowResponseConverter
 import de.jensklingenberg.ktorfit.create
 import io.ktor.client.*
 import kotlinx.coroutines.runBlocking
 
 fun main() {
 
-    val linuxKtorfit = Ktorfit(JsonPlaceHolderApi.baseUrl, HttpClient())
-    linuxKtorfit.addResponseConverter(FlowResponseConverter())
-    val api =linuxKtorfit.create<JsonPlaceHolderApi>()
+    val linuxKtorfit = Ktorfit.Builder().baseUrl(JsonPlaceHolderApi.baseUrl).httpClient(HttpClient())
+        .responseConverter(FlowResponseConverter()).build()
+
+    val api = linuxKtorfit.create<JsonPlaceHolderApi>()
     runBlocking {
-        api.getPosts().collect{
+        api.getPosts().collect {
             println(it)
         }
     }